requirements.txt
fastapi
uvicorn
tensorflow
numpy
scikit-learn
pillow
